avoid obstacles renamed to find_path and changed from spider to brain
