+++ blocks/erle-spider.js
@@ -91,15 +91,3 @@ Blockly.Blocks['spider_turn_degrees'] = {
     this.setHelpUrl('http://erlerobotics.com/docs/Robot_Operating_System/ROS/Blockly/Tutorials/Tutorial_2_Erle-Spider_Turn_degrees_block.html');
   }
 };
-
-Blockly.Blocks['spider_avoid_obstacles'] = {
-  init: function() {
-    this.appendValueInput("path_angle")
-        .appendField("Find path");
-    this.setPreviousStatement(true);
-    this.setNextStatement(true);
-    this.setColour(260);
-    this.setTooltip('');
-    this.setHelpUrl('http://www.erlerobotics.com/');
-  }
-};
